pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "digest-engine"
version = "0.1.0"
description = "Daily digest engine for hardware engineering teams using Slack-only inference"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
llm = [
    "google-genai>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
